+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ## [3.1.0] - 2020-01-17
 ### Changed
+- Added \*\*kwargs to those function parameter lists without the keyword argument capability.
 - Documentation updates.