+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - run_program:  Fixed problem with mutable default arguments issue.
 
 ### Changed
+- main:  Added program lock functionality to program.
+- main:  Added new option -y to the program.
 - crt_dump_cmd:  Changed variable name to standard convention.
 - set_db_list:  Changed variable name to standard convention.
 - run_program:  Changed variable name to standard convention.